Refactor: Organize LLM and image storage/processing into dedicated folders
