+++ src/components/mainchat/MainChat.tsx
@@ -10,7 +10,7 @@ import MessageList from './Message';
 import { messagesActions } from '../../entities/message/slice';
 import { roomsActions } from '../../entities/room/slice';
 import { Avatar, GroupChatAvatar } from '../../utils/Avatar';
-import { SendMessage, SendGroupChatMessage } from '../../services/LLMcaller';
+import { SendMessage, SendGroupChatMessage } from '../../services/llm/LLMcaller';
 import type { Sticker } from '../../entities/character/types';
 import { StickerPanel } from './StickerPanel';
 import type { FileToSend } from '../../entities/message/types';
