+++ src/services/URLs.ts
@@ -1,0 +1,6 @@
+export const GEMINI_API_BASE_URL = "https://generativelanguage.googleapis.com/v1beta/models/";
+export const VERTEX_AI_API_BASE_URL = "https://aiplatform.googleapis.com/v1/projects/{projectId}/locations/{location}/publishers/google/models/{model}:generateContent";
+export const CLAUDE_API_BASE_URL = "https://api.anthropic.com/v1/messages";
+export const OPENAI_API_BASE_URL = "https://api.openai.com/v1/chat/completions";
+export const GROK_API_BASE_URL = "https://api.x.ai/v1/chat/completions";
+export const NAI_DIFFUSION_API_BASE_URL = "https://image.novelai.net/ai/generate-image";
